requirements.txt
PyYAML>=6.0
requests>=2.28.0
beautifulsoup4>=4.11.0
feedparser>=6.0.0
python-dotenv>=1.0.0
